requirements.txt
amqp
celery
Flask
Flask-Mail
python-dotenv
redis
requests
watchdog
pylint
autopep8
Flask-SqlAlchemy
PyYAML
argh
